app01/backend/requirements.txt
Django==5.1
psycopg[binary]==3.2.1
djangorestframework==3.15.2
django-filter==24.3
dj-database-url==2.2.0
whitenoise==6.7.0
